Split arguments into separate actions plugins
