+++ src/rdiffbackup/actions/list.py
@@ -1,0 +1,38 @@
+
+from rdiffbackup.actions import BaseAction
+from rdiffbackup.utils.argopts import BooleanOptionalAction
+
+
+class ListAction(BaseAction):
+    """
+    List files at a given time, files changed since a certain time, or
+    increments, with or without size, in a given backup repository.
+    """
+    name = "list"
+
+    @classmethod
+    def add_action_subparser(cls, sub_handler):
+        subparser = super().add_action_subparser(sub_handler)
+        entity_parsers = cls._get_subparsers(
+            subparser, "entity", "files", "increments")
+        time_group = entity_parsers["files"].add_mutually_exclusive_group()
+        time_group.add_argument(
+            "--changed-since", metavar="TIME",
+            help="list files modified since given time")
+        time_group.add_argument(
+            "--at", metavar="TIME", default="now",
+            help="list files at given time (default is now/latest)")
+        entity_parsers["files"].add_argument(
+            "locations", metavar="[[USER@]SERVER::]PATH", nargs=1,
+            help="location of repository to list files from")
+        entity_parsers["increments"].add_argument(
+            "--size", action=BooleanOptionalAction, default=False,
+            help="also output size of each increment (might take longer)")
+        entity_parsers["increments"].add_argument(
+            "locations", metavar="[[USER@]SERVER::]PATH", nargs=1,
+            help="location of repository to list increments from")
+        return subparser
+
+
+def get_action_class():
+    return ListAction
