+++ src/rdiffbackup/actions/remove.py
@@ -1,0 +1,26 @@
+
+from rdiffbackup.actions import BaseAction
+
+
+class RemoveAction(BaseAction):
+    """
+    Remove the oldest increments from a backup repository.
+    """
+    name = "remove"
+
+    @classmethod
+    def add_action_subparser(cls, sub_handler):
+        subparser = super().add_action_subparser(sub_handler)
+        entity_parsers = cls._get_subparsers(
+            subparser, "entity", "increments")
+        entity_parsers["increments"].add_argument(
+            "--older-than", metavar="TIME",
+            help="remove increments older than given time")
+        entity_parsers["increments"].add_argument(
+            "locations", metavar="[[USER@]SERVER::]PATH", nargs=1,
+            help="location of repository to remove increments from")
+        return subparser
+
+
+def get_action_class():
+    return RemoveAction
