+++ src/rdiffbackup/utils/__init__.py
@@ -1,0 +1,6 @@
+"""
+In this package are modules which wrap standard or external libraries
+without being really specific to rdiff-backup (they have no backup/restore
+context), and could be replaced at any time (i.e. they don't participate
+to the API of rdiff-backup).
+"""
