+++ src/rdiffbackup/utils/argopts.py
@@ -1,0 +1,115 @@
+import argparse
+
+try:
+    from argparse import BooleanOptionalAction
+except ImportError:
+    # the class exists only since Python 3.9
+    class BooleanOptionalAction(argparse.Action):
+        def __init__(self,
+                     option_strings,
+                     dest,
+                     default=None,
+                     type=None,
+                     choices=None,
+                     required=False,
+                     help=None,
+                     metavar=None):
+
+            _option_strings = []
+            for option_string in option_strings:
+                _option_strings.append(option_string)
+
+                if option_string.startswith('--'):
+                    option_string = '--no-' + option_string[2:]
+                    _option_strings.append(option_string)
+
+            if help is not None and default is not None:
+                help += f" (default: {default})"
+
+            super().__init__(
+                option_strings=_option_strings,
+                dest=dest,
+                nargs=0,
+                default=default,
+                type=type,
+                choices=choices,
+                required=required,
+                help=help,
+                metavar=metavar)
+
+        def __call__(self, parser, namespace, values, option_string=None):
+            if option_string in self.option_strings:
+                setattr(namespace, self.dest, not option_string.startswith('--no-'))
+
+        def format_usage(self):
+            return ' | '.join(self.option_strings)
+
+
+class SelectAction(argparse.Action):
+    """
+    argparse Action class which can handle placeholder options, adding them all
+    together under one destination and keeping the same order as the arguments
+    on the command line.
+    e.g. '--exclude value1 --include-perhaps --max 10' is interpreted as
+    selections=[('exclude', value1), ('include-perhaps', True), ('max', 10)]
+    by just defining the arguments '--SELECT', '--SELECT-perhaps' and '--max'.
+    """
+
+    placeholder = 'SELECT'
+    default_dest = 'selections'
+
+    def __init__(self, option_strings, dest,
+                 type=str, nargs=None, help=None, default=None, **kwargs):
+        """
+        Initialize the placeholder-argument object, making sure that both
+        exclude and include arguments are allowed, that booleans have
+        a meaningful True value, and that all values are by default are
+        gathered under the same 'selections' destination.
+        """
+        # because the argparse framework always sets 'dest',
+        # we need to guess if dest was explicitly set, and if not,
+        # we can overwrite it with the default value
+        if ('--' + dest.replace('_', '-')) in option_strings:
+            dest = self.default_dest
+        # we want to make sure that toggles/booleans have a correct value
+        if type is bool and nargs is None:
+            nargs = 0
+            if default is None:
+                default = True
+        # replace placeholder with both include and exclude options
+        include_opts = list(map(
+            lambda x: x.replace(self.placeholder, 'include'), option_strings))
+        exclude_opts = list(map(
+            lambda x: x.replace(self.placeholder, 'exclude'), option_strings))
+        if exclude_opts != include_opts:
+            # SELECT was found hence we need to duplicate the options
+            # and update accordingly the help text
+            option_strings = exclude_opts + include_opts
+            if help:
+                help = help.replace(self.placeholder, 'exclude/include')
+                if default is None:
+                    help += " (no default)"
+                elif default:
+                    help += " (default is include)"
+                else:
+                    help += " (default is exclude)"
+        super().__init__(option_strings, dest,
+                         type=type, nargs=nargs, help=help, default=default,
+                         **kwargs)
+
+    def __call__(self, parser, namespace, values, option_string=None):
+        """
+        append the selection criteria (option_string, values) to the
+        ordered list of selection criteria.
+        """
+
+        old_list = getattr(namespace, self.dest, [])
+        # namespace is "too intelligent", it always returns None even if
+        # the parameter isn't previously defined
+        if old_list is None:
+            old_list = []
+        # append the option string and values to the selections list
+        if values == [] and self.default is not None:
+            values = self.default
+        setattr(namespace, self.dest,
+                old_list + [(option_string.replace('--', ''), values)])
