+++ testing/commontest.py
@@ -9,7 +9,7 @@ import subprocess
 from rdiff_backup.log import Log
 from rdiff_backup import Globals, Hardlink, SetConnections, Main, \
     selection, rpath, eas_acls, rorpiter, Security, hash
-from rdiffbackup import arguments
+from rdiffbackup import actions
 
 RBBin = os.fsencode(shutil.which("rdiff-backup") or "rdiff-backup")
 
@@ -174,7 +174,7 @@ def InternalBackup(source_local,
     Globals.current_time = current_time
     Globals.security_level = "override"
     Globals.set("no_compression_regexp_string",
-                os.fsencode(arguments.DEFAULT_NOT_COMPRESSED_REGEXP))
+                os.fsencode(actions.DEFAULT_NOT_COMPRESSED_REGEXP))
 
     cmdpairs = _internal_get_cmd_pairs(source_local, dest_local,
                                        src_dir, dest_dir)
@@ -227,7 +227,7 @@ def InternalRestore(mirror_local,
     Main._restore_root_set = 0
     Globals.security_level = "override"
     Globals.set("no_compression_regexp_string",
-                os.fsencode(arguments.DEFAULT_NOT_COMPRESSED_REGEXP))
+                os.fsencode(actions.DEFAULT_NOT_COMPRESSED_REGEXP))
 
     cmdpairs = _internal_get_cmd_pairs(mirror_local, dest_local,
                                        mirror_dir, dest_dir)
@@ -450,7 +450,7 @@ def BackupRestoreSeries(source_local,
     """
     Globals.set('preserve_hardlinks', compare_hardlinks)
     Globals.set("no_compression_regexp_string",
-                os.fsencode(arguments.DEFAULT_NOT_COMPRESSED_REGEXP))
+                os.fsencode(actions.DEFAULT_NOT_COMPRESSED_REGEXP))
     time = 10000
     dest_rp = rpath.RPath(Globals.local_connection, dest_dirname)
     restore_rp = rpath.RPath(Globals.local_connection, restore_dirname)
@@ -512,7 +512,7 @@ def MirrorTest(source_local,
     """Mirror each of list_of_dirnames, and compare after each"""
     Globals.set('preserve_hardlinks', compare_hardlinks)
     Globals.set("no_compression_regexp_string",
-                os.fsencode(arguments.DEFAULT_NOT_COMPRESSED_REGEXP))
+                os.fsencode(actions.DEFAULT_NOT_COMPRESSED_REGEXP))
     dest_rp = rpath.RPath(Globals.local_connection, dest_dirname)
     old_force_val = Main._force
     Main._force = 1
